refactor: test 유틸 패키지 위치 변경
