+++ src/test/java/roomescape/reservation/repository/ReservationDaoTest.java
@@ -24,7 +24,7 @@ import roomescape.reservation.domain.Name;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.Theme;
-import roomescape.reservation.utils.JdbcTemplateUtils;
+import roomescape.utils.JdbcTemplateUtils;
 
 @ActiveProfiles("test")
 @JdbcTest
