+++ src/test/java/roomescape/utils/JdbcTemplateUtils.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.utils;
+package roomescape.utils;
 
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -13,6 +13,7 @@ public final class JdbcTemplateUtils {
             jdbcTemplate.update("delete from reservation");
             jdbcTemplate.update("delete from reservation_time");
             jdbcTemplate.update("delete from theme");
+            jdbcTemplate.update("delete from member");
         } catch (DataAccessException e) {
             throw new RuntimeException("테이블 삭제 중 오류 발생", e);
         }
